Allow more parameters to be set in parameter file
The offset for the periodic boundary conditions can now be set in the
parameter files. The code can now loop over multiple stages of offset
functions. A more demanding check is done for accepting a step size. It
is also possible to start from a previous stage, and checkpoints include
both the stage and the gamma.