Generalize boundary specification
The boundary domains and conditions can now be specified in separate
parts of the parameter file. The conditions can be either dirichlet or
pair constraints, and have references to the domains they act on. Now
the parameter file can include any combination of pair constraints and
dirichlet boundary conditions. The main thing that is still not
generalized properly is the function that provides the x, y, and z for
the pair constraint boundary function. I also removed all reference to
periodic functions, and removed the modified header file.