Add function to output integrated results
These go in a separate file easily readable with analysis scripts (csv
with space delimiter). I also split up the integrate function so that
the calculation of normal and shear forces is in its own function, as is
the printing to standard out. I also made all these values be class
members.